refactor: move address-related functions
move functions for address-parsing/-resolution to a submodule.

fix #580
