+++ src/lib.rs
@@ -51,3 +51,5 @@ pub mod port_strategy;
 pub mod benchmark;
 
 pub mod scripts;
+
+pub mod address;
